Refactor, enable CORS, add more products
