+++ product-service/package.json
@@ -17,13 +17,11 @@
     "babel-preset-env": "^1.6.0",
     "jest": "^28.1.3",
     "jest-cli": "28.1.3",
+    "serverless-auto-swagger": "^2.8.0",
     "serverless-offline": "^8.8.1",
     "serverless-webpack": "^5.3.1",
     "webpack": "^4.35.2"
   },
   "author": "The serverless webpack authors (https://github.com/elastic-coders/serverless-webpack)",
-  "license": "MIT",
-  "dependencies": {
-    "serverless-auto-swagger": "^2.8.0"
-  }
+  "license": "MIT"
 }
