--- product-service/src/core/di/index.js
@@ -1,1 +1,0 @@
-export * from './container';
