+++ product-service/src/functions/get-product-list.js
@@ -1,14 +1,17 @@
-import { ProductService } from '../core';
-import { resolve } from '../core/di';
+import { resolve } from '../core';
+import { ProductService } from './services';
+import { buildHeaders } from '../helpers';
 
 export async function getProductList(event) {
+    const headers = buildHeaders();
+
     const service = resolve(ProductService);
     const products = await service.getAll();
 
     return {
+        headers,
         statusCode: 200,
         body: JSON.stringify(products),
-        headers: {},
         isBase64Encoded: false,
     };
 }
