+++ product-service/src/helpers/build-headers.js
@@ -1,0 +1,6 @@
+export function buildHeaders() {
+    return {
+        'Access-Control-Allow-Origin': '*',
+        'Access-Control-Allow-Credentials': true,
+    };
+}
