+++ product-service/src/helpers/index.js
@@ -1,0 +1,1 @@
+export * from './build-headers';
