+++ product-service/src/mocks/products.js
@@ -1,10 +1,44 @@
 export const MOCK_PRODUCTS = [
     {
-        name: 'Keychron K2',
-        price: 89
+        id: 0,
+        title: 'Anne Pro 2',
+        description: 'Wired/Wireless Mechanical Keyboard (Gateron Brown Switch/White Case) - Full Keys Programmable - True RGB Backlit - Tap Arrow Keys - Double Shot PBT Keycaps - NKRO - 1900mAh Battery',
+        price: 89,
+        count: 45,
     },
     {
-        name: 'Anne Pro 2',
+        id: 1,
+        title: 'Vortex Race 3',
+        description: 'Vortex\'s newest 75% keyboard, the 83-key Race 3! Do you need dedicated arrow keys that your 60% doesn\'t provide, but don\'t want the size of a TKL? A 75% may be just for you! Most of the keys are in the "normal" spot, so there\'s nearly no learning curve for this keyboard',
         price: 69,
+        count: 20,
+    },
+    {
+        id: 2,
+        title: 'Keychron K2',
+        description: 'System: Windows/Android/Mac/iOS; Battery: 4000mAh Rechargeable li-polymer battery; BT Working Time(RGB): Up to 10 hours (Lab test result may vary by actual use); Connection: Bluetooth 3.0 and Type-C cable',
+        price: 94,
+        count: 31,
+    },
+    {
+        id: 3,
+        title: 'Varmilo VA87M Sea Melody',
+        description: 'Not available',
+        price: 126,
+        count: 8,
+    },
+    {
+        id: 4,
+        title: 'Varmilo x MK x Ducky MIYA Pro Flare Star Wireless Special Edition',
+        description: '65%, Wireless/USB-C',
+        price: 155,
+        count: 12,
+    },
+    {
+        id: 5,
+        title: 'Varmilo Sword 2-108 Flying Tigers',
+        description: 'This is the 2nd generation of the Sword series. The case is made from CNC all-aluminum and has a four-stage color separation design that allows you to see the separation from multiple angles. It also has a detachable magnetic top that can be changed to achieve new looks. Additional colors will also be added in the future. The Sword series are the only keyboards on the market with a metal case and the Varmilo EC switches. The top mount structure boasts a gasket mount, which is a silicone strip added between the PCB at the top and bottom of the case. This strip reduces vibration when a key is pressed. The keyboard also utilizes a silicone ring between the PCB and screw holes along with a built-in sound absorbing pad to reduce vibration and pinging. These techniques are usually reserved for high end custom keyboards only.',
+        price: 472,
+        count: 15,
     },
 ];
